优化readme, 增加性能对比

+++ hash_interface.go
@@ -1,5 +1,5 @@
 package collapsar
 
 type HashInterface interface {
-	Hash(key string) uint32
+	Hash(key string) uint64
 }
